requirements.txt
torch==2.1.0
torchvision==0.16.0

ultralytics==8.1.0

# SAM - install from source, PyPI version is outdated
# Run: pip install git+https://github.com/facebookresearch/segment-anything.git

onnx==1.15.0
onnxruntime-gpu==1.17.0

fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6    # Required for FastAPI file uploads

prometheus-client==0.19.0

opencv-python-headless==4.9.0.80   # headless = no GUI, correct for servers
albumentations==1.3.1
numpy==1.26.3
scikit-learn==1.4.0

wandb==0.16.2

tqdm==4.66.1
pyyaml==6.0.1